requirements.txt
sqlalchemy
google-search-results
serpapi
requests
beautifulsoup4
google-generativeai
streamlit
pandas
python-dotenv
plotly
flask
flask-cors
